linux provisioning, tiny tweaks, tmux configuration

+++ zsh.d.symlink/aliases.zsh
@@ -1,6 +1,6 @@
-#!/bin/zsh
+alias reload!='. ~/.zshrc'
 
-if $(gls &>/dev/null) 
+if $(gls &>/dev/null)
 then
   alias ls="gls -F --color"
   alias ll="gls -lh --color"
@@ -10,8 +10,8 @@ fi
 
 alias tma="tmux attach"
 
-alias fixopenwith='/System/Library/Frameworks/CoreServices.framework/Frameworks/LaunchServices.framework/Support/lsregister -kill -r -domain local -domain system -domain user'
-alias ffdev='/Applications/Firefox.app/Contents/MacOS/firefox-bin -no-remote -P dev'
-
-alias reload!='. ~/.zshrc'
-
+# Mac-specific aliases
+if [[ $OSTYPE == darwin* ]]; then
+  alias fixopenwith='/System/Library/Frameworks/CoreServices.framework/Frameworks/LaunchServices.framework/Support/lsregister -kill -r -domain local -domain system -domain user'
+  alias ffdev='/Applications/Firefox.app/Contents/MacOS/firefox-bin -no-remote -P dev'
+fi
